Prefix environment variable to avoid conflicts
When a redis image is connected with an alias "redis", docker creates an
environment variable called REDIS_PORT and is set to a url.

More here:
https://docs.docker.com/v1.8/userguide/dockerlinks/

"Additionally, Docker creates an environment variable called
<alias>_PORT. This variable contains the URL of the source container’s
first exposed port. The ‘first’ port is defined as the exposed port with
the lowest number. For example, consider the
WEBDB_PORT=tcp://172.17.0.82:5432 variable. If that port is used for
both tcp and udp, then the tcp one is specified."