Let both stderr and stdout be streamed at the same time
There were cases where one command in the build commands would fail but
nothing was printed until the script exited because we waited until all
stdout was read before we read any of the stderr. This commit
introduces a separate thread for each stream so errors are shown as soon
as they happen (we don't wait for the command to finish).