Use an SSH key and an SSH URL for fetching repos
Switched from the https method, which uses an OAuth2 access token,
to the SSH method, which uses an SSH key, for fetching the target
repo that contains the tests the worker needs to run.

We do this to increase our security exposure as the OAuth2 tokens
won't be transmitted to the workers anymore. Instead, an SSH key
(that has to be already in place) will be used. The SSH keys can
be set up to allow access to a single repo only, as opposed to the
tokens that are tied to the repo owner and will allow full access
to any repo allowed by the token's scopes.